% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxsize_forestplot.R
\name{boxsize_forestplot}
\alias{boxsize_forestplot}
\title{Size for the boxes of forestplot}
\usage{
boxsize_forestplot(data)
}
\arguments{
\item{data}{Specificied as on the example : "O:/Codes R/Forestplot/shbg.rdata"}
}
\description{
Boxes have to be smaller for stratified analyzed (boys/girls) than total.
}
\examples{
data<-as.data.frame(c("all","girls","all","all","boys"))
colnames(data)<-"pop"
boxsize_forestplot(data)

}
