% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortest.R
\name{cortest}
\alias{cortest}
\title{Correlation graphics}
\usage{
cortest(mat)
}
\arguments{
\item{mat}{A matrix of continuous variables}
}
\description{
Build a correlation matrix for correlation plots
}
\examples{

Pierre <- c(8,9,15)
Nathalie <- c(17,12,13)
Jacques <- c(11,15,7)
Julie <- c(5,12,19)
matrice <- matrix(c(Pierre, Nathalie, Jacques, Julie), nrow=4, ncol=3, byrow=TRUE)
M<-cor(na.omit(matrice),method="spearman")
p.mat <- cortest(na.omit(matrice))
#corrplot(M, method="shade",type="upper",tl.col="black", tl.srt=45,p.mat = p.mat,
#sig.level = 0.05, insig = "blank",addCoef.col = "black",diag=F)
}
