% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist.forme.R
\name{Dist.forme}
\alias{Dist.forme}
\title{Distribution graphs}
\usage{
Dist.forme(x)
}
\arguments{
\item{x}{a vector of continuous value}
}
\description{
Graphs to check the shape of a continuous variable's distribution
}
\examples{
Dist.forme(c(14,15,18,21,20,22))

}
