% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Double.R
\name{Double}
\alias{Double}
\title{Double}
\usage{
Double(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A vector containing the Double values
}
\description{
Comput the Double value of a vector
}
\details{
The Double is calculated by multiplying each value by 2
}
\examples{
Double(runif(3))
}
