% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_sex.R
\name{forestplot_sex}
\alias{forestplot_sex}
\title{Forestplot graphs}
\usage{
forestplot_sex(data, exposure, titre)
}
\arguments{
\item{data}{Specificied as on the example : "O:/Codes R/Forestplot/shbg.rdata"}

\item{exposure}{Names indicating the explicative parameter (such as exposure indicator)}

\item{titre}{Title of the graph}
}
\value{
A forest plot
}
\description{
Print graphs by groups (in total and by sex)
}
\examples{
data<-as.data.frame(c("all","girls","all","girls","boys","boys"))
colnames(data)<-"pop"
data$Pvalue<-c(0.05,0.02,0.1,0.001,0.8,0.05)
data$Estimate<-c(1.1,1.2,2.3,1.26,1.01,1.1)
data$LowerCL<-c(1.05,1.15,0.9,1.20,0.6,1.05)
data$UpperCL<-c(1.15,1.25,3.1,1.30,1.6,1.15)
data$CI<-c("[1.05,1.15]","[1.15;1.25]","[0.9;3.1]","[1.20;1.30]","[0.6;1.6]","[1.05,1.15]")
forestplot_sex(data,c("example n°1","example n°2"),"Example")

}
