% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Mode calculation}
\usage{
getmode(v)
}
\arguments{
\item{v}{Factor}
}
\description{
Calculate the mode of a categorical variable
}
\examples{
getmode (c(112,15,18,15,415,41,5,4,44,51,51,2,65,4,8,5,5,8,45,489))

}
