% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_load_packages.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Install and load packages}
\usage{
install_load_packages(name_library)
}
\arguments{
\item{name_library}{Name of the package}
}
\description{
Automatically install a package from Lyon CRAN mirror and load it
}
\examples{
install_load_packages("installr")

}
