% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyenne.R
\name{moyenne}
\alias{moyenne}
\title{Calcul la moyenne}
\usage{
moyenne(x)
}
\arguments{
\item{x}{Un vecteur de numeriques}
}
\value{
une valeur numérique (la moyenne)
}
\description{
Une fonction qui calcule la moyenne
arithmetique d'un vecteur
}
\examples{
moyenne(1:10)
}
