% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{mydata}
\format{un dataframe
\describe{
 \item{Sepal.Length}{longueur spale.}
 \item{Sepal.Width}{largeur spale.}
 \item{Petal.Length}{longueur petale.}
 \item{Petal.Width}{largeur petale.}
 \item{Petal.Width}{espece.}
}}
\source{
copie de iris
}
\usage{
mydata
}
\description{
mydata
}
\keyword{datasets}
