% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_colors.R
\name{nice_colors}
\alias{nice_colors}
\title{Friendly colors}
\usage{
nice_colors()
}
\value{
A list of color set : raspberries, strawberries, sunday, blues, diversity, pairedcol, redNblue and bysex
}
\description{
Select nice colors for graphs
}
\examples{
nice_colors()$bysex


}
