% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_imputations.R
\name{pool_imputations}
\alias{pool_imputations}
\title{Pool imputations}
\usage{
pool_imputations(beta, sd, N)
}
\arguments{
\item{beta}{estimates from regression models}

\item{sd}{standard errors from regression models}

\item{N}{number of multiple imputations or simulations}
}
\value{
pooled beta and 95\% confidence interval
}
\description{
Pool estimators from multiple imputations
}
\examples{
beta<-c(45,42)
sd<-c(14,19)
pool_imputations(beta,sd,2)
}
