% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_LD.R
\name{simulation_LD}
\alias{simulation_LD}
\title{Imputation of zero values}
\usage{
simulation_LD(var, LD, ID, N = 75000, dist = "lognormal")
}
\arguments{
\item{var}{vector of values to impute}

\item{LD}{limit of detection}

\item{ID}{number of individuals}

\item{N}{number of random variables to be generated}

\item{dist}{shape of the distribution : lognormal (by default) or normal}
}
\description{
Impute missing values, below the limit of detection
}
\examples{
data<-c(0,45,48,65,47)
LOD<-40
simulation_LD(data,LOD,5)

}
