% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_NA.R
\name{simulation_NA}
\alias{simulation_NA}
\title{Imputation of NA}
\usage{
simulation_NA(var, LD, ID, N = 5e+06)
}
\arguments{
\item{var}{vector of values to impute}

\item{LD}{limit of detection}

\item{ID}{number of individuals}

\item{N}{number of random variables to be generated}
}
\value{
vector of values
}
\description{
Impute missing values, below the limit of detection
}
\examples{
data<-c(NA,45,48,65,47)
LOD<-40
simulation_NA(data,LOD,5)

}
