% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercile.R
\name{tercile}
\alias{tercile}
\title{Tercile}
\usage{
tercile(data, na = F)
}
\arguments{
\item{data}{A continuous variable}

\item{na}{Boolean, omit (T) or not (F) missing values}
}
\value{
data_cl, factor with 3 levels
}
\description{
Categorize a continuous variable among terciles
}
\examples{
tercile(c(112,15,18,15,415,41,5,4,44,51,51,2,65,4,8,5,5,8,45,489))

}
