% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_forestplot.R
\name{text_forestplot}
\alias{text_forestplot}
\title{Text for forestplot}
\usage{
text_forestplot(data, exposure)
}
\arguments{
\item{data}{Specificied as on the example : "O:/Codes R/Forestplot/shbg.rdata"}

\item{exposure}{Names indicating the explicative parameter (such as exposure indicator)}
}
\description{
Estimates and 95%CI in total and by sex
}
\examples{
data<-as.data.frame(c("all","girls","all","girls","boys","boys"))
colnames(data)<-"pop"
data$Pvalue<-c(0.05,0.02,0.1,0.001,0.8,0.05)
data$Estimate<-c(1.1,1.2,2.3,1.26,1.01,1.1)
data$CI<-c("[1.05,1.15]","[1.15;1.25]","[0.9;3.1]","[1.20;1.30]","[0.6;1.6]","[1.05,1.15]")
text_forestplot(data,c("example n° 1","example n°2"))

}
