% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_log_plus_c.R
\name{transform_log_plus_c}
\alias{transform_log_plus_c}
\title{Transformation log(x+c) with offset}
\usage{
transform_log_plus_c(x)
}
\arguments{
\item{x}{a vector of continuous variables}
}
\value{
log(x+best.c), vector with log-transformed variables
}
\description{
Log-tranform a variable x with a offset to prevent errors with zero values. Offset chosen to
minimize distribution skewness.
}
\examples{
transform.log.plus.c(c(15,7,5,1,8,0,85,0))

}
